Update struct definitions
# Conflicts:
#	Sources/OpenGraph_SPI/Graph/GraphDescription.mm
#	Sources/OpenGraph_SPI/include/OGGraphDescription.h

# Conflicts:
#	Sources/OpenGraphCxx/include/OpenGraph/OGAttributeType.h
#	Sources/OpenGraphCxx/include/OpenGraph/OGClosure.h
#	Sources/OpenGraphCxx/include/OpenGraph/OGValue.h

+++ Sources/OpenGraphCxx/include/OpenGraph/OGAttributeFlags.h
@@ -10,7 +10,7 @@
 
 #include <OpenGraph/OGBase.h>
 
-typedef OG_OPTIONS(uint32_t, OGAttributeFlags) {
+typedef OG_OPTIONS(uint8_t, OGAttributeFlags) {
     OGAttributeFlagsDefault = 0,
     OGAttributeFlagsActive = 1 << 0,
     OGAttributeFlagsRemovable = 1 << 1,
